refactor: extract Operator and Operand to use in other class
