+++ src/main/java/Operands.java
@@ -12,17 +12,7 @@ public class Operands {
                 .collect(Collectors.toCollection(LinkedList::new));
     }
 
-
-    private static class Operand {
-        private final int num;
-
-        public Operand(String str) {
-            this.num = Integer.parseInt(str);
-        }
-
-        @Override
-        public String toString() {
-            return String.valueOf(num);
-        }
+    public Operand poll() {
+        return this.operands.poll();
     }
 }
